Push rebuild so that CVE-2016-6323 is tagged in security feed
We had to make changes to ypkg-gen-history as it wasn't emitting the security
tag on CVE updates, meaning users were unaware that the updates they were
getting (i.e. previous glibc build, and the recent x11 libs), were in fact
security updates.

The Solus Software Center specifically uses the tag to correctly categorize
the updates, thus it is essential the tagging mechanism is working correctly.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>