Workaround 32-bit Aspyr ports not working anymore with GCC 7
This temporary fix is to workaround an upstream issue in GCC that produces
unaligned code when compiling gcc, which in turn breaks 32-bit Aspyr ports
such as Civilisation and Borderlands 2.

For now we'll use `-mstackrealign` when building the pure 32-bit variant,
and improve our runtime Snap to mitigate issues with performance (if any)
in the alt-AVX2 builds.

This update will be synced as an emergency shannon update.

 - https://gcc.gnu.org/bugzilla/show_bug.cgi?id=81646 (closed dupe)
 - https://gcc.gnu.org/bugzilla/show_bug.cgi?id=40838 (new since 2009.)

This is also built against rebootstrapped toolchain and modified GCC
which now contains two patches, one of which to help with SSE2 alignment.

Important: This regression was not fixed in GCC 7.2 as previously thought
but rather seems to be an unfortunate consequence of cross-compiling our
libraries for `-m32` on a native 64bit toolchain.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>